[soony.kwonn] 방탈출 step1,2 리뷰 요청 드립니다. (#25)
* docs: 개발 요구사항 정리

* build: JWT, lombok 의존성 추가

* refactor: lombok 적용

* refactor: java 컨벤션 적용

* feat: login 기능 추가

* feat: login API 추가

* refactor: request, response 파일명에 Dto 접미사 추가

* feat: MemberResponseDto 생성

* feat: findByUsername() 생성

* feat: findUsernameByToken() 생성

* feat: /members/me API 및 테스트 구현

* docs: 내 정보 조회 기능 구체화

* feat: 멤버 username 중복 처리

* feat: GlobalExceptionHandler 생성

* feat: EmptyResultDataAccessException 처리

* docs: step2 기능 명세

* build: jib 설정

* build: 배포를 위한 k8s yaml 생성

* feat: NoAccessTokenException, JwtException 예외처리

* refactor: lombok, java 컨벤션 적용

* feat: AuthenticationPrincipalArgumentResolver를 이용하여 username 추출

* feat: reservation 생성시 AuthenticationPrincipalArgumentResolver 활용

* feat: reservation 삭제시 권한 확인

* feat: UnauthorizedException 처리

* refactor: exception 패키지 분리

* test: reservationService.deleteById() 테스트

* test: reservationController.deleteReservation() 테스트

* feat: 존재하지 않는 회원정보일 때 로그인 실패

* refactor: java 컨벤션 적용

* docs: README.md 수정

* test: 예외 별 테스트 메서드 구체화

* refactor: 테스트 메서드 포맷 통일

* docs: 미션 명세 url 추가

* feat: Repository findByUsernameAndPasswordTest 메서드 추가

* feat: TokenRequestDto를 통해 유효한 회원인지 확인 기능 추가

* refactor: 메서드 이름 및 기능 변경

* refactor: 예외 클래스 변경

* feat: login시 비밀번호까지 확인하도록 변경

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -2,7 +2,9 @@ package nextstep.reservation;
 
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.DuplicateEntityException;
+import nextstep.support.exception.DuplicateEntityException;
+import nextstep.support.exception.NotExistEntityException;
+import nextstep.support.exception.UnauthorizedException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -21,7 +23,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, String username) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,8 +36,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
-        );
+                username);
 
         return reservationDao.save(newReservation);
     }
@@ -49,10 +50,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, String username) {
         Reservation reservation = reservationDao.findById(id);
+
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new NotExistEntityException("예약번호가 유효하지 않습니다.");
+        }
+        if (reservation.getName() != username) {
+            throw new UnauthorizedException("자신의 예약이 아닙니다.");
         }
 
         reservationDao.deleteById(id);
